+++ ruma-events/CHANGELOG.md
@@ -3,6 +3,9 @@
 Breaking changes:
 
 * Update strum dependency to 0.19
+* Use `ruma_identifiers::{ServerName, ServerKeyId}` in `signatures` fields of
+  `pdu::RoomV1Pdu, RoomV1PduStub, RoomV3Pdu, RoomV3PduStub}` and
+  `room::member::SignedContent`.
 
 # 0.22.0